Hide token in log and error reading
The token is sensitive data, so it's hidden in the logfile now.

2020.08.06 13:36:24 4: GoEChargerCloud (GoECharger) - Send with URI: https://api.go-e.co/api_status?token=*hidden*

2020.08.06 13:38:13 4: GoEChargerCloud (GoECharger) - Send with URI: https://api.go-e.co/api?token=*hidden*&payload=alw=0

Besides that, the token should be replaced in the error reading, too. This is untested so far...

Http_lastRequestError https://api.go-e.co/api_status?token=xxxx: Can't connect(2) to https://api.go-e.co:443: SSL connect attempt failed